# fix(sd): resolve benchmark hanging and improve buffer processing performance