--- a/firmware/src/services/sd_card_services/sd_card_manager.h
+++ b/firmware/src/services/sd_card_services/sd_card_manager.h
@@ -11,7 +11,7 @@
 #include "Util/CircularBuffer.h"
 
 #define SD_CARD_MANAGER_CONF_RBUFFER_SIZE 512
-#define SD_CARD_MANAGER_CONF_WBUFFER_SIZE 5120
+#define SD_CARD_MANAGER_CONF_WBUFFER_SIZE 5120  // Keep at 5KB to fit within uint16_t circular buffer limit
 
 #define SD_CARD_MANAGER_CONF_DIR_NAME_LEN_MAX 40
 #define SD_CARD_MANAGER_CONF_FILE_NAME_LEN_MAX 40