## Summary
Fixes SD card benchmark hanging issue and implements initial performance optimizations.

## Problem
SD card benchmarks hang indefinitely when attempting writes ≥275KB due to circular buffer filling up faster than the SD task can drain it.

## Changes
1. **Add timeout protection to WriteToBuffer**
   - 2-second timeout when waiting for buffer space
   - Returns 0 (failure) if timeout occurs
   - Prevents infinite loops

2. **Optimize buffer processing**
   - Reduce SD task delay from 5ms to 1ms
   - Process up to 4 chunks (20KB) per cycle vs single 5KB chunk
   
3. **Code quality improvements**
   - Replace magic numbers with named constants
   - Add proper error messages for debugging

## Test Results
| Size | Before | After | Status |
|------|--------|-------|--------|
| 275KB | HUNG | 174 KB/s | ✅ Fixed |
| 500KB | HUNG | 186 KB/s | ✅ Fixed |
| 1MB | HUNG | 190 KB/s | ✅ Fixed |

## Testing
- Tested with Kootion 32GB UHS-I Class 10 SD card
- All benchmarks complete successfully
- Consistent ~180-190 KB/s throughput
- No impact on other system functions

## Related Issues
- Fixes #73
- Progress towards #72 (SD performance optimization tracking)

## Next Steps
Further optimizations tracked in #72:
- SPI clock speed optimization
- DMA implementation
- Larger buffer architecture